// web/projects/em/src/app/settings/schedule/model/move-task-folder-request.ts
/*
 * This file is part of StyleBI.
 * Copyright (C) 2024  InetSoft Technology
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
import { ScheduleTaskModel } from "../../../../../../shared/schedule/model/schedule-task-model";
import { RepositoryTreeNode } from "../../content/repository/repository-tree-node";

export interface MoveTaskFolderRequest {
   target: RepositoryTreeNode;
   tasks: ScheduleTaskModel[];
   folders: string[];
}